Server sends /now responses with microseconds since epoch
Millisecond and microsecond is fractional part of float response.

Also removed search.service.js from frontend codebase.

+++ backend/moonstream/data.py
@@ -57,7 +57,7 @@ class NowResponse(BaseModel):
     Schema for responses on /now endpoint
     """
 
-    epoch_time: int
+    epoch_time: float
 
 
 class SubscriptionUpdate(BaseModel):
